리플레이 시 라인 가속도 및 버튼 구현 클래스 분할

+++ guestbook/Pen_Str.h
@@ -20,7 +20,7 @@ typedef struct Pen_Info
     LPARAM      Coordinate;     // ÁÂÇ¥
     int         Width;          // Ææ ±½±â
     COLORREF    Color;          // Ææ »ö±ò
-    __int64     Time;           // ½Ã°£
+    DWORD       Time;           // ½Ã°£
     UINT        State;          // »óÅÂ (LBUTTON)
 } PEN_INFO;
